refactor: extract maximum key length functionality into dedicated trait
